refactor: move oyw-related components to a single folder, create smaller components and a TeamLeads type
