+++ src/i18n/es/index.ts
@@ -1,5 +1,4 @@
 import { type I18nData } from '../type'
-import { OurStory } from './OurStory'
 import { Ai } from './Ai'
 import { blogDetails } from './BlogDetails'
 import { blogListing } from './BlogListing'
@@ -11,14 +10,16 @@ import { home } from './Home'
 import { JobDescription } from './JobDescription'
 import { JobListing } from './JobListing'
 import { navbar } from './Navbar'
+import { OneYoungWorld } from './OneYoungWorld'
 import { OurPeople1 } from './OurPeople1'
 import { OurPeople2 } from './OurPeople2'
 import { ourResearch } from './OurResearch'
+import { OurStory } from './OurStory'
 import { OurSupporters } from './OurSupporters'
 import { Publications } from './publications'
-import { tellYourStory } from './TellYourStory'
 import { ShareYourCough } from './ShareYourCough'
-import { OneYoungWorld } from './OneYoungWorld'
+import { TeamLeads } from './TeamLeads'
+import { tellYourStory } from './TellYourStory'
 
 export const data: I18nData = {
   locale: 'es',
@@ -42,4 +43,5 @@ export const data: I18nData = {
   footer,
   ShareYourCough,
   OneYoungWorld,
+  TeamLeads,
 }
