+++ src/i18n/types/TeamLeads.ts
@@ -1,0 +1,13 @@
+import { type StaticImageData } from 'next/image'
+
+export interface TeamLeads {
+  title: string
+  cards: TeamLeadCard[]
+}
+
+export interface TeamLeadCard {
+  name: string
+  texts: string[]
+  image: StaticImageData
+  altText: string
+}
